ref: optimized Dockerfile structure
Moved the package installation into a separate stage to make use of the caching mechanism. This way the packages are cached as long as the package.json is not modified.

By having updated the package-checksums on the local machine the yarn installation now does not require clearing the cache and updating the checksums on every execution.